// app/src/generated/types/asset.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

/** What the invest share of a claim is swapped into. */
export enum Asset {
  Sol,
  Btc,
  Eth,
}

export type AssetArgs = Asset;

export function getAssetEncoder(): FixedSizeEncoder<AssetArgs> {
  return getEnumEncoder(Asset);
}

export function getAssetDecoder(): FixedSizeDecoder<Asset> {
  return getEnumDecoder(Asset);
}

export function getAssetCodec(): FixedSizeCodec<AssetArgs, Asset> {
  return combineCodec(getAssetEncoder(), getAssetDecoder());
}